fix for mac movement issue
* updating only elan-interface-forwarding-entries when mac is learned
from vlan provider port which is already learned on VM port.
* updating MACs in elan ds and dpn when MAC is actually moved from VM port
to another VM port.
* handling MAC learning from VLAN provider port if it is a newly learned
mac

+++ vpnservice/elanmanager/elanmanager-impl/src/main/java/org/opendaylight/netvirt/elan/internal/ElanSmacFlowEventListener.java
@@ -86,20 +86,26 @@ public class ElanSmacFlowEventListener implements SalFlowListener {
             }
             MacEntry macEntry = elanUtils.getInterfaceMacEntriesOperationalDataPath(interfaceName, physAddress);
             InterfaceInfo interfaceInfo = interfaceManager.getInterfaceInfo(interfaceName);
+            String elanInstanceName = elanTagInfo.getName();
             if (macEntry != null && interfaceInfo != null) {
                 WriteTransaction deleteFlowTx = broker.newWriteOnlyTransaction();
-                elanUtils.deleteMacFlows(ElanUtils.getElanInstanceByName(broker, elanTagInfo.getName()), interfaceInfo,
+                elanUtils.deleteMacFlows(ElanUtils.getElanInstanceByName(broker, elanInstanceName), interfaceInfo,
                         macEntry, deleteFlowTx);
                 ListenableFuture<Void> result = deleteFlowTx.submit();
                 addCallBack(result, srcMacAddress);
             }
             InstanceIdentifier<MacEntry> macEntryIdForElanInterface = ElanUtils
                     .getInterfaceMacEntriesIdentifierOperationalDataPath(interfaceName, physAddress);
-            InstanceIdentifier<MacEntry> macEntryIdForElanInstance = ElanUtils
-                    .getMacEntryOperationalDataPath(elanTagInfo.getName(), physAddress);
             WriteTransaction tx = broker.newWriteOnlyTransaction();
             tx.delete(LogicalDatastoreType.OPERATIONAL, macEntryIdForElanInterface);
-            tx.delete(LogicalDatastoreType.OPERATIONAL, macEntryIdForElanInstance);
+            MacEntry macEntryInElanInstance = elanUtils.getMacEntryForElanInstance(elanInstanceName,
+                    physAddress).orNull();
+            if (macEntryInElanInstance != null
+                    && macEntryInElanInstance.getInterface().equals(interfaceName)) {
+                InstanceIdentifier<MacEntry> macEntryIdForElanInstance = ElanUtils
+                        .getMacEntryOperationalDataPath(elanInstanceName, physAddress);
+                tx.delete(LogicalDatastoreType.OPERATIONAL, macEntryIdForElanInstance);
+            }
             ListenableFuture<Void> writeResult = tx.submit();
             addCallBack(writeResult, srcMacAddress);
         }
